LPS-44962 Rename
