+++ modules/portal/portal-sso-cas/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Liferay Portal CAS SSO
+Bundle-Name: Liferay Portal SSO CAS
 Bundle-SymbolicName: com.liferay.portal.sso.cas
 Bundle-Version: 1.0.0
 Import-Package:\
