+++ modules/portal/portal-sso-ntlm/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Liferay Portal NTLM SSO
+Bundle-Name: Liferay Portal SSO NTLM
 Bundle-SymbolicName: com.liferay.portal.sso.ntlm
 Bundle-Version: 1.0.0
 Import-Package:\
