+++ modules/portal/portal-sso-token/bnd.bnd
@@ -1,11 +1,11 @@
 Bundle-Description: Implmentation of Liferay Portal SSO integration \
 	that works with token based SSO solutions like Shibboleth, \
 	OAM, Siteminder
-Bundle-Name: Liferay Portal Token Based SSO
-Bundle-SymbolicName: com.liferay.portal.sso.tokenbased
+Bundle-Name: Liferay Portal SSO Token
+Bundle-SymbolicName: com.liferay.portal.sso.token
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.liferay.portal.sso.tokenbased.spi.*;version=${plugin.full.version}
+	com.liferay.portal.sso.token.spi.*;version=${plugin.full.version}
 Include-Resource:\
 	classes,\
 	@../../../lib/portal/bnd.jar!/aQute/bnd/annotation/metatype/*
