+++ modules/portal/portal-sso-token/src/com/liferay/portal/sso/token/auth/RequestHeaderTokenRetriever.java
@@ -12,11 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.sso.tokenbased.auth;
+package com.liferay.portal.sso.token.auth;
 
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.sso.tokenbased.spi.TokenLocation;
-import com.liferay.portal.sso.tokenbased.spi.TokenRetriever;
+import com.liferay.portal.sso.token.spi.TokenLocation;
+import com.liferay.portal.sso.token.spi.TokenRetriever;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -26,18 +25,18 @@ import org.osgi.service.component.annotations.Component;
  * @author Michael C. Han
  */
 @Component(immediate = true, service = TokenRetriever.class)
-public class RequestTokenRetriever implements TokenRetriever {
+public class RequestHeaderTokenRetriever implements TokenRetriever {
 
 	@Override
 	public String getLoginToken(
 		HttpServletRequest request, String userTokenName) {
 
-		return ParamUtil.getString(request, userTokenName);
+		return request.getHeader(userTokenName);
 	}
 
 	@Override
 	public TokenLocation getTokenLocation() {
-		return TokenLocation.REQUEST;
+		return TokenLocation.REQUEST_HEADER;
 	}
 
 }
