+++ modules/portal/portal-sso-token/src/com/liferay/portal/sso/token/auth/RequestTokenRetriever.java
@@ -12,13 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.sso.tokenbased.auth;
+package com.liferay.portal.sso.token.auth;
 
-import com.liferay.portal.sso.tokenbased.spi.TokenLocation;
-import com.liferay.portal.sso.tokenbased.spi.TokenRetriever;
+import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.sso.token.spi.TokenLocation;
+import com.liferay.portal.sso.token.spi.TokenRetriever;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -26,20 +26,18 @@ import org.osgi.service.component.annotations.Component;
  * @author Michael C. Han
  */
 @Component(immediate = true, service = TokenRetriever.class)
-public class SessionTokenRetriever implements TokenRetriever {
+public class RequestTokenRetriever implements TokenRetriever {
 
 	@Override
 	public String getLoginToken(
 		HttpServletRequest request, String userTokenName) {
 
-		HttpSession httpSession = request.getSession();
-
-		return (String)httpSession.getAttribute(userTokenName);
+		return ParamUtil.getString(request, userTokenName);
 	}
 
 	@Override
 	public TokenLocation getTokenLocation() {
-		return TokenLocation.SESSION;
+		return TokenLocation.REQUEST;
 	}
 
 }
