+++ modules/portal/portal-sso-token/src/com/liferay/portal/sso/token/spi/TokenRetriever.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.sso.tokenbased.spi;
+package com.liferay.portal.sso.token.spi;
 
 import javax.servlet.http.HttpServletRequest;
