--- modules/portal/portal-sso-tokenbased/src/com/liferay/portal/sso/tokenbased/auth/RequestHeaderTokenRetriever.java
@@ -1,42 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.sso.tokenbased.auth;
-
-import com.liferay.portal.sso.tokenbased.spi.TokenLocation;
-import com.liferay.portal.sso.tokenbased.spi.TokenRetriever;
-
-import javax.servlet.http.HttpServletRequest;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Michael C. Han
- */
-@Component(immediate = true, service = TokenRetriever.class)
-public class RequestHeaderTokenRetriever implements TokenRetriever {
-
-	@Override
-	public String getLoginToken(
-		HttpServletRequest request, String userTokenName) {
-
-		return request.getHeader(userTokenName);
-	}
-
-	@Override
-	public TokenLocation getTokenLocation() {
-		return TokenLocation.REQUEST_HEADER;
-	}
-
-}
